back/requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.2.0
appnope==0.1.4
arel==0.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-asgi-testclient==1.4.11
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blis==1.0.1
CacheControl==0.14.1
cacheout==0.14.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
cymem==2.0.10
dask==2024.11.2
dataframe_image==0.2.6
dataset-tools==0.1.4
debugpy==1.8.7
DecisionTreeClassifier==0.0.7
decorator==5.1.1
defusedxml==0.7.1
distinctipy==1.3.4
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
fastai==2.7.18
fastapi==0.109.0
fastcore==1.7.22
fastdownload==0.0.7
fastjsonschema==2.21.0
fastprogress==1.0.3
fcm-django==2.2.1
ffmpeg-python==0.2.0
filelock==3.16.1

flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
geographiclib==2.0
geojson==3.1.0
geopy==2.4.1
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
giturlparse==0.12.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gorilla-cli==0.0.10
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
h5py==3.12.1
halo==0.0.31
hpack==4.0.0
html2image==2.0.5
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
imagesize==1.4.1
imutils==0.5.4
inflect==7.4.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.6.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
log-symbols==0.0.14
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.11
namex==0.0.8
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
np==1.0.2
numerize==0.12
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==24.1
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.5
psutil==5.9.8
psycopg2-binary==2.9.10
PTable==0.9.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydicom==2.4.4
Pygments==2.18.0
PyJWT==2.9.0
Pympler==1.0.1
pynrrd==0.4.3
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==6.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.3
robotframework==7.2.2
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.7.1
rpds-py==0.21.0
rsa==4.9
scikit-elm==0.21a0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.31.0
setuptools==72.2.0
shapely==2.0.2
shellingham==1.5.4
SimpleITK==2.4.0
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spinners==0.0.24
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
StandardScaler==0.5
starlette==0.35.1
stringcase==1.2.0
stripe==11.2.0
supervisely==6.73.242
swapper==1.4.0
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
titlecase==2.4.1
toolz==1.0.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.5.0
trio==0.30.0
trio-websocket==0.12.2
typeguard==4.4.1
typer==0.15.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.47
ultralytics-thop==2.0.12
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.32.1
uvloop==0.21.0
varname==0.13.5
wasabi==1.1.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
whitenoise==6.9.0
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.18.0
zstd==1.5.5.1
